Misplaced bullet point as introduced by P0138R2
The bullet point 3.9 ("Otherwise, if T is an enumeration with a fixed underlying type…") is not reachable, as all corresponding cases would be covered by 3.7 ("Otherwise, if the initializer list has a single element of type E…"). Moreover, 3.8 and 3.9 cover disjoint scenarios. Therefore 3.9 is moved to a more appropriate location.